\name{cjshmm-package}
\alias{cjshmm-package}
\alias{cjshmm}
\docType{package}
\title{
\packageTitle{cjshmm}
}
\description{
\packageDescription{cjshmm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cjshmm}
\packageIndices{cjshmm}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{cjshmm}

Maintainer: \packageMaintainer{cjshmm}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
