\name{CJSRL.hmm.adapt.mcmc}
\alias{CJSRL.hmm.adapt.mcmc}
\title{Implements survival analysis with detectability and dead recovery designed for grassland winter survival dataset.}
\description{Implements survival analysis with detectability and dead recovery designed for grassland winter survival dataset. Put functions in a package because this was the solution suggested to get parallelization to work with C code compiled with Rcpp.}
\author{
Mevin Hooten
}
