\name{CJSRL.hmm.PML.pred}
\alias{CJSRL.hmm.PML.pred}
\title{Implements penalized maximum likelihood portion of survival analysis.}
\description{Implements survival analysis with detectability and dead recovery designed for grassland winter survival dataset. Put functions in a package because this was the solution suggested to get parallelization to work with C code compiled with Rcpp.}
\author{
Mevin Hooten
}
